requirements.txt
absl-py==0.11.0
addict==2.4.0
antspyx==0.2.7
argon2-cffi==20.1.0
astor==0.8.1
astunparse==1.6.3
async-generator==1.10
attrs==20.3.0
backcall==0.2.0
bleach==3.3.0
cached-property==1.5.2
cachetools==4.2.2
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
chart-studio==1.1.0
clang==5.0
cupy-cuda100==9.0.0
cycler==0.10.0
dataclasses==0.8
decorator==4.4.2
defusedxml==0.7.1
dicom2nifti==2.2.12
dill==0.3.4
entrypoints==0.3
et-xmlfile==1.0.1
fastrlock==0.6
flatbuffers==1.12
future==0.18.2
gast==0.2.2
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googleapis-common-protos==1.53.0
grpcio==1.40.0
h5py==2.10.0
idna==2.10
imageio==2.9.0
importlib-metadata==3.4.0
importlib-resources==5.2.2
ipydatawidgets==4.2.0
ipykernel==5.5.3
ipython==7.16.1
ipython-genutils==0.2.0
ipyvolume==0.5.2
ipywebrtc==0.6.0
ipywidgets==7.6.3
jedi==0.18.0
Jinja2==2.11.3
joblib==1.0.0
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
keras-tqdm==1.0.1
kiwisolver==1.3.1
llvmlite==0.36.0
Markdown==3.3.3
MarkupSafe==1.1.1
matplotlib==3.3.4
MedPy==0.4.0
mistune==0.8.4
mock==4.0.3
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
networkx==2.5
neurite==0.1
nibabel==3.2.1
nilearn==0.7.0
notebook==6.3.0
numpy==1.19.5
oauthlib==3.1.1
open3d==0.12.0
OpenEXR==1.3.2
opt-einsum==3.3.0
packaging==20.8
pandas==1.1.5
pandocfilters==1.4.3
parso==0.8.2
patsy==0.5.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.1.0
pkg_resources==0.0.0
plotly==4.14.3
plyfile==0.7.3
probreg==0.3.1
prometheus-client==0.10.0
promise==2.3
prompt-toolkit==3.0.18
protobuf==3.14.0
psutil==5.8.0
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybind11==2.6.2
pycparser==2.20
pycpd==2.0.0
pydicom==2.1.2
Pygments==2.8.1
pynrrd==0.4.2
pyparsing==2.4.7
pyrsistent==0.17.3
pystrum==0.1
python-dateutil==2.8.1
python-utils==2.5.6
pythreejs==2.3.0
pytz==2021.1
PyWavelets==1.1.1
PyYAML==5.4.1
pyzmq==22.0.3
requests==2.25.1
requests-oauthlib==1.3.0
retrying==1.3.3
rsa==4.7.2
scikit-image==0.17.2
scikit-learn==0.24.1
scipy==1.5.4
Send2Trash==1.5.0
SimpleITK==2.0.2
six==1.15.0
sklearn==0.0
statsmodels==0.12.2
tabulate==0.8.9
tensorboard==1.14.0
tensorboard-data-server==0.6.1
tensorflow-addons==0.14.0
tensorflow-datasets==4.4.0
tensorflow-estimator==1.14.0
tensorflow-gpu==1.14.0
tensorflow-metadata==1.2.0
termcolor==1.1.0
terminado==0.9.4
testpath==0.4.4
tf-keras-contrib==2.0.8
threadpoolctl==2.1.0
tifffile==2020.9.3
tikzplotlib==0.9.7
tornado==6.1
tqdm==4.56.0
traitlets==4.3.3
traittypes==0.2.1
transformations==2020.1.1
trimesh==3.9.29
typeguard==2.12.1
typing-extensions==3.7.4.3
urllib3==1.26.3
voxelmorph==0.1
vtk==9.0.1
wcwidth==0.2.5
webcolors==1.11.1
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
xlrd==2.0.1
xlwt==1.3.0
zipp==3.4.0
